requirements.txt
streamlit
openai
python-dotenv
plotly
pandas
numpy 